requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0